Refactor CLI into modular commands
